Chunk size long

+++ jhove-modules/src/main/java/edu/harvard/hul/ois/jhove/module/GifModule.java
@@ -306,7 +306,7 @@ public class GifModule extends ModuleBase
              * left, so it will get checksummed. */
             for (;;) {
                 try {
-                    int n = skipBytes (_dstream, 2048, this);
+                    long n = skipBytes (_dstream, 2048, this);
                     if (n == 0) {
                         break;
                     }
